+++ app/engine/core/updatable/node.ts
@@ -1,112 +1,7 @@
 import * as term from "~/engine/term";
 import { type Updatable, type PublishableState } from "~/engine/interfaces";
 import { Task } from "./task";
-import { error } from "~/engine";
-
-/**
- * 노드 내 CPU 코어를 표현하는 클래스
- */
-export class Core implements Updatable {
-  #id: term.Identifier;
-  #currentTask: Task | null;
-
-  private constructor(id: term.Identifier, currentTask: Task | null) {
-    this.#id = id;
-    this.#currentTask = currentTask;
-  }
-
-  private clone(
-    update: Partial<{
-      id: term.Identifier;
-      currentTask: Task | null;
-    }> = {}
-  ): Core {
-    return new Core(
-      update.id ?? this.#id,
-      update.currentTask !== undefined ? update.currentTask : this.#currentTask
-    );
-  }
-
-  public static idle(): Core {
-    return new Core(new term.Identifier("core"), null);
-  }
-
-  /**
-   * 코어에 작업 할당
-   * @param task 할당할 작업
-   */
-  public assignTask(task: Task): Core {
-    if (this.isBusy()) throw new error.BusyCoreError("Core is busy");
-
-    return this.clone({ currentTask: task });
-  }
-
-  /**
-   * 코어에서 작업 해제
-   */
-  public releaseTask(): Core {
-    if (!this.isBusy()) throw new error.NoTaskError("No task assigned");
-
-    return this.clone({ currentTask: null });
-  }
-
-  /**
-   * 코어 상태 업데이트
-   * @param deltaTime 경과 시간
-   */
-  public after(deltaTime: term.MilliSecond): Core {
-    if (!this.#currentTask) return this.clone();
-
-    const updatedTask = this.#currentTask.after(deltaTime);
-
-    if (updatedTask.isCompleted()) {
-      return this.releaseTask();
-    }
-    return this.clone({ currentTask: updatedTask });
-  }
-
-  /**
-   * 코어 초기화
-   */
-  public reset(): Core {
-    if (this.#currentTask) {
-      return this.clone({ currentTask: this.#currentTask.reset() });
-    }
-    return this.clone({ currentTask: null });
-  }
-
-  /**
-   * 코어 ID 가져오기
-   */
-  public id(): string {
-    return this.#id.toString();
-  }
-
-  /**
-   * 코어 사용 중 여부 확인
-   */
-  public isBusy(): boolean {
-    return this.currentTask() !== null;
-  }
-
-  /**
-   * 코어의 현재 작업 가져오기
-   */
-  public currentTask(): Task | null {
-    return this.#currentTask;
-  }
-
-  /**
-   * 코어 상태 객체 반환
-   */
-  public state(): any {
-    return {
-      id: this.id(),
-      busy: this.isBusy(),
-      task: this.#currentTask ? this.#currentTask.state() : null,
-    };
-  }
-}
+import { Core } from "./core";
 
 /**
  * 다중 코어를 가진 노드를 표현하는 클래스
@@ -122,7 +17,7 @@ export class Node implements Updatable {
    * @param coreCount 코어 개수
    * @param position 노드 위치
    */
-  private constructor(
+  constructor(
     id: term.Identifier,
     cores: Core[],
     taskQueue: Task[],
@@ -154,13 +49,13 @@ export class Node implements Updatable {
    * 새 노드 인스턴스 생성
    * @param coreCount 코어 개수
    */
-  public static boot(coreCount: number): Node {
+  public static boot<T extends Node>(coreCount: number): T {
     return new Node(
       new term.Identifier("node"),
       Array.from({ length: coreCount }, () => Core.idle()),
       [],
       new term.Position(0, 0)
-    );
+    ) as T;
   }
 
   /**
@@ -177,8 +72,8 @@ export class Node implements Updatable {
    * 작업 대기열에 작업 등록
    * @param task 등록할 작업
    */
-  public registerTask(task: Task): Node {
-    return this.clone({ taskQueue: [...this.#taskQueue, task] });
+  public registerTask<T extends Node>(task: Task): T {
+    return this.clone({ taskQueue: [...this.#taskQueue, task] }) as T;
   }
 
   /**
